add captive magnet voids for cups
kludge in cylindrical voids to the cup tree. customizer to enable "print-in-magnets" but no other options. the void starts at z=0.4, which is supposed to be two (bottom) layers.

I think it would be a good idea to set a 'lh' global because some features (like the material between the magnet void and the outer surface of the object) are best calibrated to multiples to the layer height.

I did this work originally on commit 8d511b6b. rebasing took a few hours.