add captive magnet voids for cups (#183)
* add captive magnet voids for cups

kludge in cylindrical voids to the cup tree. customizer to enable "print-in-magnets" but no other options. the void starts at z=0.4, which is supposed to be two (bottom) layers.

I think it would be a good idea to set a 'lh' global because some features (like the material between the magnet void and the outer surface of the object) are best calibrated to multiples to the layer height.

I did this work originally on commit 8d511b6b. rebasing took a few hours.

* update calculateFloorHeight and change primary captive magnet setting from boolean to number

i realized that instead of setting some kind of global layerheight somehwere, we can just give the user the power to set the height. that's what you've done with everything else. that makes more sense and is more flexible

then i updated calculateFloorHeight by adding the setting  to depth calculations. then i added the setting to the parameters of EfficientFloorAttachmentCaps

these changes give decent behavior with various combinations of magnet_captive_height, efficient_floor, and floor_thickness that i tested.